# Conflicts:
#	build.gradle
#	src/main/kotlin/ru/hits/bdui/expression/ExpressionEvaluator.kt

+++ src/main/kotlin/ru/hits/bdui/domain/screen/components/properties/Size.kt
@@ -16,7 +16,7 @@ sealed interface Size {
      * @property fraction значение в диапазоне от 0 до 1 (включительно)
      */
     data class Weighted(
-        val fraction: Int
+        val fraction: Double
     ) : Size {
         override val type: String = "weighted"
     }
